requirements.txt
#python 3.12
langchain
langchain_openai
langchain_community
langchain_chroma
langgraph
gradio
chromadb
pydantic
python-dotenv
beautifulsoup4
tiktoken
requests
sentence-transformers